Winter Term Document Revisions Part 1
Updated Abstract, Problem Statement, Requirements Document, and
Technology Review to use new document formatting. Simplified large
portions of the Latex code so that the documents are much easier to
read, and smaller. We still need to revise some of the documents I
listed above, such as updating the Gantt chart within the Requirements
Document.