Various: Toasts, Editbook / newbook
Working on edit book / newbook, as these pages are not working properly.
New book works 80% correct, with just the renaming of the newly created
chapter causing problems.

Edit book, allows for the rearrangement of chapters, or to add a chapter
from search, but problems arise when trying to add a chapter from my
chapters, as dragula isn't being invoked correctly.

I created mychapterside for this purpose.